feat: expose jsonschema.Schema instead of interface{}
when trying to integrate with ollama i needed to do quite a lot of data
conversion (ollama representation of a tool is very different from what
we use here). i found it easier to make this data conversion if i deal
directly with a jsonschema.schema instead of an interface{}.

on my tests here everything is still working with this change. if there
is a better way of reading a jsonschema.schema from the tool.inputschema
please let me know.