Update Client.ListTools
The MCP specification defines the cursor parameter as an optional
string. In the world of TypeScript, an optional string is either string
or undefined. It does not include null values.

The Client.ListTools function allows for null values to be passed in,
which is not aligned with the official specifications and therefore does
not play well with some MCP servers.

References:

- MCP official definition: https://github.com/modelcontextprotocol/modelcontextprotocol/blob/ec1418927730e3ffeede8ea6475d6f4c36f6d10c/schema/draft/schema.ts#L368
- Optional parameters in TypeScript: https://www.typescriptlang.org/docs/handbook/2/functions.html#optional-parameters

Fixes https://github.com/metoro-io/mcp-golang/issues/125